+++ tests/InfluxDB/dashboard_tdg_static_with_instance_variable.sh
@@ -1,7 +1,0 @@
-#!/bin/sh
-
-make DATASOURCE=influxdb \
-     POLICY=autogen \
-     MEASUREMENT=tarantool_http \
-     WITH_INSTANCE_VARIABLE=TRUE \
-     build-static-tdg-influxdb