+++ CHANGELOG.md
@@ -11,6 +11,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
   It affects all custom-built dashboards which had used `cluster` (#224)
 - Renamed build `replication` panels section to `replication_cartridge`.
   It affects all custom-built dashboards which had used `replication` (#224)
+- Remove deprecated `build-static-*` targets from `make` (#224)
 
 ### Fixed
 - Removed extra quotation mark for TDG jobs_average_panel query (#220)